Workshop: correct filename case
