+++ packages/user/Workshop/ui/src/components/control-panel.tsx
@@ -5,14 +5,14 @@ import {
     Status,
     appMetadataQueryKey,
     useAppMetadata,
-} from "@/hooks/useAppMetadata";
-import { useCurrentApp } from "@/hooks/useCurrentApp";
-import { useDeleteCsp } from "@/hooks/useDeleteCsp";
-import { usePublishApp } from "@/hooks/usePublishApp";
-import { useSetCacheMode } from "@/hooks/useSetCacheMode";
-import { useSetCsp } from "@/hooks/useSetCsp";
-import { SiteConfigResponse, useSiteConfig } from "@/hooks/useSiteConfig";
-import { useSpa } from "@/hooks/useSpa";
+} from "@/hooks/use-app-metadata";
+import { useCurrentApp } from "@/hooks/use-current-app";
+import { useDeleteCsp } from "@/hooks/use-delete-csps";
+import { usePublishApp } from "@/hooks/use-publish-app";
+import { useSetCacheMode } from "@/hooks/use-set-cache-mode";
+import { useSetCsp } from "@/hooks/use-set-csps";
+import { SiteConfigResponse, useSiteConfig } from "@/hooks/use-site-config";
+import { useSpa } from "@/hooks/use-spa";
 
 import { queryClient } from "@shared/lib/queryClient";
 import { zAccount } from "@shared/lib/schemas/account";
@@ -20,7 +20,7 @@ import { Button } from "@shared/shadcn/ui/button";
 import { Label } from "@shared/shadcn/ui/label";
 import { toast } from "@shared/shadcn/ui/sonner";
 
-import { CheckCard } from "./Check-Card";
+import { CheckCard } from "./check-card";
 import { CspForm } from "./csp-form";
 import { ServiceUpload } from "./service-upload";
