+++ packages/user/Workshop/ui/src/hooks/use-set-csps.ts
@@ -6,7 +6,7 @@ import { zAccount } from "@shared/lib/schemas/account";
 import { supervisor } from "@shared/lib/supervisor";
 import { toast } from "@shared/shadcn/ui/sonner";
 
-import { SiteConfigResponse, siteConfigQueryKey } from "./useSiteConfig";
+import { SiteConfigResponse, siteConfigQueryKey } from "./use-site-config";
 
 const Params = z.object({
     account: zAccount,
